+++ src/puuid/base.py
@@ -1,7 +1,7 @@
 """
-pUUID Base Implementation.
+pUUID base implementation.
 
-Provides the abstract base class and version-specific implementations for Prefixed UUIDs.
+Provides the abstract base class and version-specific implementations for prefixed UUIDs.
 """
 
 import annotationlib
@@ -17,7 +17,6 @@ from typing import (
     final,
     get_args,
     get_origin,
-    overload,
     override,
 )
 from uuid import NAMESPACE_DNS, UUID, uuid1, uuid3, uuid4, uuid5, uuid6, uuid7, uuid8
@@ -50,17 +49,9 @@ else:
 @final
 class ERR_MSG:
     UUID_VERSION_MISMATCH = "Expected 'UUID' with version '{expected}', got '{actual}'"
-    FACTORY_UNSUPPORTED = "'PUUID.factory' is only supported for 'PUUIDv1', 'PUUIDv4', 'PUUIDv6', 'PUUIDv7' and 'PUUIDv8'!"
     PREFIX_DESERIALIZATION_ERROR = "Unable to deserialize prefix '{prefix}', separator '_' or UUID for '{classname}' from '{serial_puuid}'!"
     INVALID_TYPE_FOR_SERIAL_PUUID = "'{classname}' can not be created from invalid type '{type}' with value '{value}'!"
     EMPTY_PREFIX_DISALLOWED = "Empty prefix is not allowed for '{classname}'!"
-    INVALID_PUUIDv1_ARGS = "Invalid 'PUUIDv1' arguments: Provide either 'node' and 'clock_seq' or a 'uuid'!"
-    INVALID_PUUIDv3_ARGS = "Invalid 'PUUIDv3' arguments: Provide either 'namespace' and 'name' or a 'uuid'!"
-    INVALID_PUUIDv5_ARGS = "Invalid 'PUUIDv5' arguments: Provide either 'namespace' and 'name' or a 'uuid'!"
-    INVALID_PUUIDv6_ARGS = "Invalid 'PUUIDv6' arguments: Provide either 'node' and 'clock_seq' or a 'uuid'!"
-    INVALID_PUUIDv8_ARGS = (
-        "Invalid 'PUUIDv8' arguments: Provide either 'a', 'b' and 'c' or 'uuid'!"
-    )
 
 
 class PUUIDError(Exception):
@@ -113,7 +104,7 @@ class PUUIDBase[TPrefix: str](ABC):
     """Abstract Generic Base Class for Prefixed UUIDs."""
 
     _prefix: ClassVar[str] = ""
-    _serial: str | None
+    _serial: str | None = None
     _uuid: UUID
 
     def __init_subclass__(cls, **kwargs: Any) -> None:
@@ -140,7 +131,7 @@ class PUUIDBase[TPrefix: str](ABC):
         )
 
     @abstractmethod
-    def __init__(self, *, uuid: UUID) -> None: ...
+    def __init__(self, uuid: UUID, *, check_version: bool = True) -> None: ...
 
     def __new__(cls, *args: Any, **kwargs: Any) -> Self:
         instance = super().__new__(cls)
@@ -186,32 +177,12 @@ class PUUIDBase[TPrefix: str](ABC):
         str
             The formatted string (e.g., `<prefix>_<uuid-hex-string>`).
         """
-        cached = self._serial
-        if cached is not None:
-            return cached
+        if self._serial is not None:
+            return self._serial
 
         serial = self._format_serial()
         self._serial = serial
         return serial
-
-    @classmethod
-    def factory(cls) -> Self:
-        """
-        Create a new instance using default generation.
-
-        Supported by version variants that allow generation without arguments.
-
-        Returns
-        -------
-        Self
-            A new instance of the pUUID class.
-
-        Raises
-        ------
-        PUUIDError
-            If the variant does not support parameterless generation.
-        """
-        raise PUUIDError(ERR_MSG.FACTORY_UNSUPPORTED)
 
     @classmethod
     def from_string(cls, serial_puuid: str) -> Self:
@@ -360,68 +331,58 @@ class PUUIDBase[TPrefix: str](ABC):
 class PUUIDv1[TPrefix: str](PUUIDBase[TPrefix]):
     """Prefixed UUID Version 1 (MAC address and time)."""
 
-    _uuid: UUID
-    _serial: str | None
     VERSION: Final[int] = 1
-
-    @overload
-    def __init__(
-        self, *, node: int | None = None, clock_seq: int | None = None
-    ) -> None: ...
-
-    @overload
-    def __init__(self, *, uuid: UUID) -> None: ...
 
     def __init__(
         self,
+        uuid: UUID,
         *,
-        node: int | None = None,
-        clock_seq: int | None = None,
-        uuid: UUID | None = None,
+        check_version: bool = True,
     ) -> None:
         """
         Initialize a PUUIDv1.
 
         Parameters
         ----------
-        node : int | None, optional
-            Hardware address. If None, `uuid1` generates a random value.
-        clock_seq : int | None, optional
-            Clock sequence.
-        uuid : UUID | None, optional
-            Existing UUID v1 instance.
+        uuid : UUID
+            An UUIDv1 instance.
 
         Raises
         ------
         PUUIDError
-            If arguments are inconsistent or the UUID version is incorrect.
+            If the UUID version is incorrect.
         """
-        match node, clock_seq, uuid:
-            case int() | None, int() | None, None:
-                self._uuid = uuid1(node, clock_seq)
-            case None, None, UUID(version=1):
-                self._uuid = uuid
-            case None, None, UUID(version=version):
-                raise PUUIDError(
-                    ERR_MSG.UUID_VERSION_MISMATCH.format(expected=1, actual=version)
-                )
-            case _:
-                raise PUUIDError(ERR_MSG.INVALID_PUUIDv1_ARGS)
 
-        self._serial = None
+        if check_version and (uuid.version != 1):
+            raise PUUIDError(
+                ERR_MSG.UUID_VERSION_MISMATCH.format(expected=1, actual=uuid.version)
+            )
+        self._uuid = uuid
 
-    @override
     @classmethod
-    def factory(cls) -> Self:
+    def factory(
+        cls,
+        *,
+        node: int | None = None,
+        clock_seq: int | None = None,
+    ) -> Self:
         """
-        Create a new PUUIDv1 instance using current time and MAC address.
+        Create a new PUUIDv1 instance. using current time and MAC address.
+
+        Parameters
+        ----------
+        node : int | None, optional
+            MAC address.
+        clock_seq : int | None, optional
+            The current time.
 
         Returns
         -------
         Self
-            A new pUUID v1 instance.
+            A new PUUIDv1 instance.
         """
-        return cls()
+
+        return cls(uuid1(node, clock_seq), check_version=False)
 
 
 ################################################################################
@@ -432,25 +393,38 @@ class PUUIDv1[TPrefix: str](PUUIDBase[TPrefix]):
 class PUUIDv3[TPrefix: str](PUUIDBase[TPrefix]):
     """Prefixed UUID Version 3 (MD5 hash of namespace and name)."""
 
-    _uuid: UUID
-    _serial: str | None
     VERSION: Final[int] = 3
-
-    @overload
-    def __init__(self, *, namespace: UUID, name: str | bytes) -> None: ...
-
-    @overload
-    def __init__(self, *, uuid: UUID) -> None: ...
 
     def __init__(
         self,
+        uuid: UUID,
         *,
-        namespace: UUID | None = None,
-        name: str | bytes | None = None,
-        uuid: UUID | None = None,
+        check_version: bool = True,
     ) -> None:
         """
         Initialize a PUUIDv3.
+
+        Parameters
+        ----------
+        uuid : UUID
+            An UUIDv3 instance.
+
+        Raises
+        ------
+        PUUIDError
+            If the UUID version is incorrect.
+        """
+
+        if check_version and (uuid.version != 3):
+            raise PUUIDError(
+                ERR_MSG.UUID_VERSION_MISMATCH.format(expected=3, actual=uuid.version)
+            )
+        self._uuid = uuid
+
+    @classmethod
+    def factory(cls, *, namespace: UUID, name: str | bytes) -> Self:
+        """
+        Create a new PUUIDv3.
 
         Parameters
         ----------
@@ -458,27 +432,14 @@ class PUUIDv3[TPrefix: str](PUUIDBase[TPrefix]):
             Namespace UUID.
         name : str | bytes | None, optional
             The name used for hashing.
-        uuid : UUID | None, optional
-            Existing UUID v3 instance.
 
-        Raises
-        ------
-        PUUIDError
-            If arguments are inconsistent or the UUID version is incorrect.
+        Returns
+        -------
+        Self
+            A new PUUIDv3 instance.
         """
-        match namespace, name, uuid:
-            case UUID(), str() | bytes(), None:
-                self._uuid = uuid3(namespace, name)
-            case None, None, UUID(version=3):
-                self._uuid = uuid
-            case None, None, UUID(version=version):
-                raise PUUIDError(
-                    ERR_MSG.UUID_VERSION_MISMATCH.format(expected=3, actual=version)
-                )
-            case _:
-                raise PUUIDError(ERR_MSG.INVALID_PUUIDv3_ARGS)
 
-        self._serial = None
+        return cls(uuid3(namespace, name), check_version=False)
 
 
 ################################################################################
@@ -489,32 +450,34 @@ class PUUIDv3[TPrefix: str](PUUIDBase[TPrefix]):
 class PUUIDv4[TPrefix: str](PUUIDBase[TPrefix]):
     """Prefixed UUID Version 4 (randomly generated)."""
 
-    _uuid: UUID
-    _serial: str | None
     VERSION: Final[int] = 4
 
-    def __init__(self, *, uuid: UUID | None = None) -> None:
+    def __init__(
+        self,
+        uuid: UUID,
+        *,
+        check_version: bool = True,
+    ) -> None:
         """
         Initialize a PUUIDv4.
 
         Parameters
         ----------
-        uuid : UUID | None, optional
-            Existing UUID v4 instance. If None, a new random UUID is generated.
+        uuid : UUID
+            An UUIDv4 instance.
 
         Raises
         ------
         PUUIDError
-            If the provided UUID is not version 4.
+            If the UUID version is incorrect.
         """
-        if uuid is not None and uuid.version != 4:
+
+        if check_version and (uuid.version != 4):
             raise PUUIDError(
                 ERR_MSG.UUID_VERSION_MISMATCH.format(expected=4, actual=uuid.version)
             )
-        self._uuid = uuid if uuid else uuid4()
-        self._serial = None
+        self._uuid = uuid
 
-    @override
     @classmethod
     def factory(cls) -> Self:
         """
@@ -523,9 +486,10 @@ class PUUIDv4[TPrefix: str](PUUIDBase[TPrefix]):
         Returns
         -------
         Self
-            A new pUUID v4 instance.
+            A new PUUIDv4 instance.
         """
-        return cls()
+
+        return cls(uuid4(), check_version=False)
 
 
 ################################################################################
@@ -536,53 +500,58 @@ class PUUIDv4[TPrefix: str](PUUIDBase[TPrefix]):
 class PUUIDv5[TPrefix: str](PUUIDBase[TPrefix]):
     """Prefixed UUID Version 5 (SHA-1 hash of namespace and name)."""
 
-    _uuid: UUID
-    _serial: str | None
     VERSION: Final[int] = 5
-
-    @overload
-    def __init__(self, *, namespace: UUID, name: str | bytes) -> None: ...
-
-    @overload
-    def __init__(self, *, uuid: UUID) -> None: ...
 
     def __init__(
         self,
+        uuid: UUID,
         *,
-        namespace: UUID | None = None,
-        name: str | bytes | None = None,
-        uuid: UUID | None = None,
+        check_version: bool = True,
     ) -> None:
         """
         Initialize a PUUIDv5.
 
         Parameters
         ----------
-        namespace : UUID | None, optional
-            Namespace UUID.
-        name : str | bytes | None, optional
-            The name used for hashing.
-        uuid : UUID | None, optional
-            Existing UUID v5 instance.
+        uuid : UUID
+            Existing UUIDv5 instance.
 
         Raises
         ------
         PUUIDError
-            If arguments are inconsistent or the UUID version is incorrect.
+            If the UUID version is incorrect.
         """
-        match namespace, name, uuid:
-            case UUID(), str() | bytes(), None:
-                self._uuid = uuid5(namespace, name)
-            case None, None, UUID(version=5):
-                self._uuid = uuid
-            case None, None, UUID(version=version):
-                raise PUUIDError(
-                    ERR_MSG.UUID_VERSION_MISMATCH.format(expected=5, actual=version)
-                )
-            case _:
-                raise PUUIDError(ERR_MSG.INVALID_PUUIDv5_ARGS)
 
-        self._serial = None
+        if check_version and (uuid.version != 5):
+            raise PUUIDError(
+                ERR_MSG.UUID_VERSION_MISMATCH.format(expected=5, actual=uuid.version)
+            )
+        self._uuid = uuid
+
+    @classmethod
+    def factory(
+        cls,
+        *,
+        namespace: UUID,
+        name: str | bytes,
+    ) -> Self:
+        """
+        Create a new PUUIDv5 instance using random generation.
+
+        Parameters
+        ----------
+        namespace : UUID
+            Namespace UUID.
+        name : str | bytes
+            The name used for hashing.
+
+        Returns
+        -------
+        Self
+            A new PUUIDv5 instance.
+        """
+
+        return cls(uuid5(namespace=namespace, name=name), check_version=False)
 
 
 ################################################################################
@@ -593,68 +562,58 @@ class PUUIDv5[TPrefix: str](PUUIDBase[TPrefix]):
 class PUUIDv6[TPrefix: str](PUUIDBase[TPrefix]):
     """Prefixed UUID Version 6 (reordered v1 for DB locality)."""
 
-    _uuid: UUID
-    _serial: str | None
     VERSION: Final[int] = 6
-
-    @overload
-    def __init__(
-        self, *, node: int | None = None, clock_seq: int | None = None
-    ) -> None: ...
-
-    @overload
-    def __init__(self, *, uuid: UUID) -> None: ...
 
     def __init__(
         self,
+        uuid: UUID,
         *,
-        node: int | None = None,
-        clock_seq: int | None = None,
-        uuid: UUID | None = None,
+        check_version: bool = True,
     ) -> None:
         """
         Initialize a PUUIDv6.
 
         Parameters
         ----------
-        node : int | None, optional
-            Hardware address.
-        clock_seq : int | None, optional
-            Clock sequence.
-        uuid : UUID | None, optional
-            Existing UUID v6 instance.
+        uuid : UUID
+            An UUIDv6 instance.
 
         Raises
         ------
         PUUIDError
-            If arguments are inconsistent or the UUID version is incorrect.
+            If the UUID version is incorrect.
         """
-        match node, clock_seq, uuid:
-            case int() | None, int() | None, None:
-                self._uuid = uuid6(node, clock_seq)
-            case None, None, UUID(version=6):
-                self._uuid = uuid
-            case None, None, UUID(version=version):
-                raise PUUIDError(
-                    ERR_MSG.UUID_VERSION_MISMATCH.format(expected=6, actual=version)
-                )
-            case _:
-                raise PUUIDError(ERR_MSG.INVALID_PUUIDv6_ARGS)
 
-        self._serial = None
+        if check_version and (uuid.version != 6):
+            raise PUUIDError(
+                ERR_MSG.UUID_VERSION_MISMATCH.format(expected=6, actual=uuid.version)
+            )
+        self._uuid = uuid
 
-    @override
     @classmethod
-    def factory(cls) -> Self:
+    def factory(
+        cls,
+        *,
+        node: int | None = None,
+        clock_seq: int | None = None,
+    ) -> Self:
         """
-        Create a new PUUIDv6 instance using reordered time-based generation.
+        Create a new PUUIDv6 instance. using current time and MAC address.
+
+        Parameters
+        ----------
+        node : int | None, optional
+            MAC address.
+        clock_seq : int | None, optional
+            The current time.
 
         Returns
         -------
         Self
-            A new pUUID v6 instance optimized for DB locality.
+            A new PUUIDv6 instance.
         """
-        return cls()
+
+        return cls(uuid6(node, clock_seq), check_version=False)
 
 
 ################################################################################
@@ -665,43 +624,46 @@ class PUUIDv6[TPrefix: str](PUUIDBase[TPrefix]):
 class PUUIDv7[TPrefix: str](PUUIDBase[TPrefix]):
     """Prefixed UUID Version 7 (time-ordered)."""
 
-    _uuid: UUID
-    _serial: str | None
     VERSION: Final[int] = 7
 
-    def __init__(self, uuid: UUID | None = None) -> None:
+    def __init__(
+        self,
+        uuid: UUID,
+        *,
+        check_version: bool = True,
+    ) -> None:
         """
         Initialize a PUUIDv7.
 
         Parameters
         ----------
-        uuid : UUID | None, optional
-            Existing UUID v7 instance. If None, a new time-ordered UUID is generated.
+        uuid : UUID
+            An UUIDv7 instance.
 
         Raises
         ------
         PUUIDError
-            If the provided UUID is not version 7.
+            If the UUID version is incorrect.
         """
-        if uuid is not None and uuid.version != 7:
+
+        if check_version and (uuid.version != 7):
             raise PUUIDError(
                 ERR_MSG.UUID_VERSION_MISMATCH.format(expected=7, actual=uuid.version)
             )
-        self._uuid = uuid if uuid else uuid7()
-        self._serial = None
+        self._uuid = uuid
 
-    @override
     @classmethod
     def factory(cls) -> Self:
         """
-        Create a new PUUIDv7 instance using time-ordered generation.
+        Create a new PUUIDv7 instance using random generation.
 
         Returns
         -------
         Self
-            A new pUUID v7 instance.
+            A new PUUIDv7 instance.
         """
-        return cls()
+
+        return cls(uuid7(), check_version=False)
 
 
 ################################################################################
@@ -712,28 +674,44 @@ class PUUIDv7[TPrefix: str](PUUIDBase[TPrefix]):
 class PUUIDv8[TPrefix: str](PUUIDBase[TPrefix]):
     """Prefixed UUID Version 8 (custom implementation)."""
 
-    _uuid: UUID
-    _serial: str | None
     VERSION: Final[int] = 8
-
-    @overload
-    def __init__(
-        self, *, a: int | None = None, b: int | None = None, c: int | None = None
-    ) -> None: ...
-
-    @overload
-    def __init__(self, *, uuid: UUID) -> None: ...
 
     def __init__(
         self,
+        uuid: UUID,
+        *,
+        check_version: bool = True,
+    ) -> None:
+        """
+        Initialize a PUUIDv8.
+
+        Parameters
+        ----------
+        uuid : UUID
+            An UUIDv8 instance.
+
+        Raises
+        ------
+        PUUIDError
+            If the UUID version is incorrect.
+        """
+
+        if check_version and (uuid.version != 8):
+            raise PUUIDError(
+                ERR_MSG.UUID_VERSION_MISMATCH.format(expected=8, actual=uuid.version)
+            )
+        self._uuid = uuid
+
+    @classmethod
+    def factory(
+        cls,
         *,
         a: int | None = None,
         b: int | None = None,
         c: int | None = None,
-        uuid: UUID | None = None,
-    ) -> None:
+    ) -> Self:
         """
-        Initialize a PUUIDv8.
+        Create a new PUUIDv8 instance using custom generation.
 
         Parameters
         ----------
@@ -743,37 +721,11 @@ class PUUIDv8[TPrefix: str](PUUIDBase[TPrefix]):
             Second custom 12-bit value.
         c : int | None, optional
             Third custom 62-bit value.
-        uuid : UUID | None, optional
-            Existing UUID v8 instance.
-
-        Raises
-        ------
-        PUUIDError
-            If arguments are inconsistent or the UUID version is incorrect.
-        """
-        match a, b, c, uuid:
-            case int() | None, int() | None, int() | None, None:
-                self._uuid = uuid8(a, b, c)
-            case None, None, None, UUID(version=8):
-                self._uuid = uuid
-            case None, None, None, UUID(version=version):
-                raise PUUIDError(
-                    ERR_MSG.UUID_VERSION_MISMATCH.format(expected=8, actual=version)
-                )
-            case _:
-                raise PUUIDError(ERR_MSG.INVALID_PUUIDv8_ARGS)
-
-        self._serial = None
-
-    @override
-    @classmethod
-    def factory(cls) -> Self:
-        """
-        Create a new PUUIDv8 instance using custom generation.
 
         Returns
         -------
         Self
-            A new pUUID v8 instance.
+            A new PUUIDv8 instance.
         """
-        return cls()
+
+        return cls(uuid8(a, b, c), check_version=False)